requirements.txt
streamlit
opencv-python
numpy
qiskit
qiskit-aer
Pillow
scikit-image
matplotlib
